Next VC++ errors fixed

+++ companion/src/eeprominterface.cpp
@@ -1639,7 +1639,7 @@ void unregisterFirmwares()
 
 void ShowEepromErrors(QWidget *parent, const QString &title, const QString &mainMessage, unsigned long errorsFound)
 {
-  std::bitset<NUM_ERRORS> errors(errorsFound);
+  std::bitset<NUM_ERRORS> errors((unsigned long long)errorsFound);
   QStringList errorsList;
 
   errorsList << QT_TRANSLATE_NOOP("EepromInterface", "Possible causes for this:");
@@ -1674,7 +1674,7 @@ void ShowEepromErrors(QWidget *parent, const QString &title, const QString &main
 
 void ShowEepromWarnings(QWidget *parent, const QString &title, unsigned long errorsFound)
 {
-  std::bitset<NUM_ERRORS> errors(errorsFound);
+  std::bitset<NUM_ERRORS> errors((unsigned long long)errorsFound);
   QStringList warningsList;
   if (errors.test(WARNING_WRONG_FIRMWARE)) { warningsList << QT_TRANSLATE_NOOP("EepromInterface", "- Your radio probably uses a wrong firmware,\n eeprom size is 4096 but only the first 2048 are used"); }
   if (errors.test(OLD_VERSION)) { warningsList << QT_TRANSLATE_NOOP("EepromInterface", "- Your eeprom is from an old version of OpenTX, upgrading!\n You should 'save as' to keep the old file as a backup."); }
